+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ### Unreleased
 
-- Add RowsOfFieldsWithMetadata
-- Add NumericCellRenderer
+- Add RowsOfFieldsWithMetadata: allows commands to return an object with metadata that shows up in yaml/json (& etc.) formats, but is not shown in table/csv (& etc.).
+- Add NumericCellRenderer: allows commands to attach a renderer that will right-justify and add commas to numbers in a column.
 
 ### 3.1.13 - 29 November 2017